+++ AdditiveSynth/CollectorExamples/HarmonicCollectorSine.h
@@ -15,11 +15,11 @@ namespace tmf
     class HarmonicCollectorSine : public AdditiveSynthHarmonicCollector
     {
     public:
-        void collectHarmonics (juce::AudioBuffer<float>& audioBuffer, int tableSize) override
+        void collectHarmonics (juce::AudioBuffer<float>& audioBuffer, size_t tableSize) override
         {
             jassert (sampleRate > 0);
             jassert (numChannels > 0);
-            std::vector<float> table = std::vector<float>(tableSize, 0);
+            std::vector<float> table = std::vector<float> (tableSize, 0);
             if (tableSize >= 2)
             {
                 table[2] = 1;
